server/src/useCases/submitFeedbackUseCase.spec.ts
import { SubmitFeedbackUseCase } from "./submitFeedbackUseCase";

const createFeedbackSpy = jest.fn();
const sendMailSpy = jest.fn();

const submitFeedback = new SubmitFeedbackUseCase(
  { create: createFeedbackSpy },
  { sendMail: sendMailSpy }
)

describe('Submit feedback', () => {
  
  it('should be able to submit a feedback', async () => {
    await expect(submitFeedback.execute({
      type: 'BUG',
      comment: 'Example test comment',
      screenshot: 'data:image/png;base64,2223232131'
    })).resolves.not.toThrow();

    expect(createFeedbackSpy).toHaveBeenCalled();
    expect(createFeedbackSpy).toHaveBeenCalled();
  })

  it('should not be able to submit a feedback without a type', async () => {
    await expect(submitFeedback.execute({
      type: '',
      comment: 'Example test comment',
      screenshot: 'data:image/png;base64,2223232131'
    })).rejects.toThrow();
  })

  it('should not be able to submit a feedback without a comment', async () => {
    await expect(submitFeedback.execute({
      type: 'BUG',
      comment: '',
      screenshot: 'data:image/png;base64,2223232131'
    })).rejects.toThrow();
  })

  it('should not be able to submit a feedback wit an invalid screenshot', async () => {
    await expect(submitFeedback.execute({
      type: 'BUG',
      comment: 'Example test comment',
      screenshot: 'pacoca.jpg'
    })).rejects.toThrow();
  })
})